Rework pre-Glibc 2.35 versions.
