+++ manifest/x86_64/g/glibc_lib235.filelist
